requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
python-dotenv==1.0.1
django-cors-headers==4.4.0
requests==2.32.3
drf-yasg==1.21.11
psycopg2-binary==2.9.10